API: Take AsRef<Path> and Send+Sync errors
Having errors be `Send+Sync` is a good practice in the Rust ecosystem as
it allows at least passing the errors in and out of threads as it is
propagated.

`AsRef<Path>` allows more convenient use of the constructor wherein you
have the convenience of being able to pass in e.g. a string literal.