summarize: Don't panic on broken pipe
I hit #31 while piping summarize to head. The original issue was in part
due to `prettytable` panicking. That's since been upgraded to a version
with the fix, but some of the `println!`s that happen after
`table.printstd()` will also panic from the broken pipe:

  $ summarize summarize ... | head
  ...
  thread 'main' panicked at library/std/src/io/stdio.rs:1165:9:
  failed printing to stdout: Broken pipe (os error 32)
  note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace

To fix this, change the `println!`s to `writeln!`s, and ignore any
errors.

Fixes #31.